Add files via upload
                                           Glovo Dashboard Insight Story. 
DATA BASE CREATION.
The data for this visualization was obtained via web search. Each of the vendors name was obtained and an excel data base was generated with the following columns: - Customers with online presence, customers with social media handles, phone numbers and their customer ratings.
TRANSFORMATION AND LOADING.
With the use of power query, the data was cleaned to remove any errors on the created data base. 
It was then closed and apply. Going in to the visualization pain, count of total vendors was obtained, count vendors with online presence and social media was also shown on our visualization pane. 
INSIGHTS.
From our dashboard, it was shown that 14% of our the vendors on our database have website, while 88% have social media presence. The reasons why we have vendors with website on that low percentage  is as result of the fact that some of those vendors are operating on a small scale. It is easier for them to have social media account like facebook with little or no cost as against owing a website. This dashboard also shows vendors with the highest ratings.