Add files via upload
his analysis shows the sales activities of a superstore.
I performed data extraction process  to obtain data from  kaggle site. . 
STEP 2
 After data extraction, I used pd.read_csv to import the file from the part to begin data cleaning and preparation.. 
STEP 3 
Data Cleaning and Preparation I converted the data to the appropriate data type, renaming of column where necesaary and checking for null values. 
I use the head() and the info() method to view the obtained data sets both from the repository and the scrapped data. 
 STEP 4
From the dashboard, the west has highest sales records making the work performance to be high in that region. Also shown on the dashboard is the technology supplies  with the highest shipment order. There is an increase in sales from January to December with an increase in profit also. West has high sales records with technology leading the chart in the supply and with  standard segment as the highest.