Spoker Bike Sales Dashboad
This is a powerBi dashboard  showing the sales and profit records of a company which sells bike. My report shows the city that generates the highest bike purchases as a result of population. Also showing is the monthly sales trend. Ages 40-50 records the highest purchases this is as a result of highest working demography that falls in that age range. Also in my report is the bike product with the highest sales. The highest with the highest sales record was also shown. The higher the sales record the more the profit generated.