Add files via upload
The USA has the highest revenue due to high purchasing power and of course high population.
Since the US has the highest sales revenue, this invariably means they have the highest units of goods demanded or high units of sales.
The second quarter has the highest revenue income with the months of  April, May and June having a consistent revenue. This is as a results of seasonal purchases or ordering of the said goods or products.